[C-2273] Fix gigantic audio coin
