+++ packages/mobile/src/screens/app-drawer-screen/left-nav-drawer/LeftNavDrawer.tsx
@@ -1,18 +1,13 @@
-import { useCallback, useContext } from 'react'
-
 import type { User } from '@audius/common'
 import {
   FeatureFlags,
   StringKeys,
-  formatWei,
-  walletSelectors,
   accountSelectors,
-  useSelectTierInfo,
   useAccountHasClaimableRewards
 } from '@audius/common'
 import type { DrawerContentComponentProps } from '@react-navigation/drawer'
 import { DrawerContentScrollView } from '@react-navigation/drawer'
-import { TouchableOpacity, View } from 'react-native'
+import { View } from 'react-native'
 import { useSelector } from 'react-redux'
 
 import IconCrown from 'app/assets/images/iconCrown.svg'
@@ -21,21 +16,16 @@ import IconMessage from 'app/assets/images/iconMessage.svg'
 import IconSettings from 'app/assets/images/iconSettings.svg'
 import IconUpload from 'app/assets/images/iconUpload.svg'
 import IconUser from 'app/assets/images/iconUser.svg'
-import { IconAudioBadge } from 'app/components/audio-rewards'
-import { Text } from 'app/components/core'
-import { ProfilePicture } from 'app/components/user'
-import UserBadges from 'app/components/user-badges'
 import { useFeatureFlag, useRemoteVar } from 'app/hooks/useRemoteConfig'
 import { makeStyles } from 'app/styles'
 import { spacing } from 'app/styles/spacing'
 
-import { AppDrawerContext, AppDrawerContextProvider } from '../AppDrawerContext'
-import { useAppDrawerNavigation } from '../useAppDrawerNavigation'
+import { AppDrawerContextProvider } from '../AppDrawerContext'
 
+import { AccountDetails } from './AccountDetails'
 import { LeftNavLink } from './LeftNavLink'
 import { VanityMetrics } from './VanityMetrics'
 const { getAccountUser } = accountSelectors
-const { getAccountTotalBalance } = walletSelectors
 
 const messages = {
   profile: 'Profile',
@@ -51,42 +41,6 @@ type AccountDrawerProps = DrawerContentComponentProps & {
 }
 
 const useStyles = makeStyles(({ spacing, palette }) => ({
-  header: {
-    paddingLeft: spacing(4),
-    flexDirection: 'row',
-    alignItems: 'flex-start',
-    justifyContent: 'space-between'
-  },
-  accountImage: {
-    height: spacing(12.5),
-    width: spacing(12.5),
-    marginBottom: spacing(3),
-    borderWidth: 1
-  },
-  accountName: {
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-    paddingRight: spacing(2)
-  },
-  name: {
-    flexShrink: 1
-  },
-  accountBadges: {
-    flexGrow: 1,
-    alignSelf: 'center'
-  },
-  handle: {
-    paddingRight: spacing(3)
-  },
-  tokens: {
-    position: 'absolute',
-    flexDirection: 'row',
-    alignItems: 'center',
-    right: spacing(4)
-  },
-  tokenBadge: {
-    marginRight: spacing(1)
-  },
   notificationBubble: {
     height: spacing(3),
     width: spacing(3),
@@ -100,6 +54,7 @@ export const LeftNavDrawer = (props: AccountDrawerProps) => {
   const { navigation: drawerHelpers, ...other } = props
   const accountUser = useSelector(getAccountUser) as User
   if (!accountUser) return null
+
   return (
     <AppDrawerContextProvider drawerHelpers={drawerHelpers} {...other}>
       <WrappedLeftNavDrawer />
@@ -108,65 +63,14 @@ export const LeftNavDrawer = (props: AccountDrawerProps) => {
 }
 
 const WrappedLeftNavDrawer = () => {
-  const { drawerHelpers } = useContext(AppDrawerContext)
   const styles = useStyles()
-  const accountUser = useSelector(getAccountUser) as User
-  const { user_id, name, handle } = accountUser
-  const { tier } = useSelectTierInfo(user_id)
-  const totalBalance = useSelector(getAccountTotalBalance)
   const challengeRewardIds = useRemoteVar(StringKeys.CHALLENGE_REWARD_IDS)
   const hasClaimableRewards = useAccountHasClaimableRewards(challengeRewardIds)
   const { isEnabled: isChatEnabled } = useFeatureFlag(FeatureFlags.CHAT_ENABLED)
 
-  const navigation = useAppDrawerNavigation()
-
-  const handlePressAccount = useCallback(() => {
-    navigation.push('Profile', { handle: 'accountUser' })
-    drawerHelpers.closeDrawer()
-  }, [navigation, drawerHelpers])
-
-  const handlePressRewards = useCallback(() => {
-    navigation.push('AudioScreen')
-    drawerHelpers.closeDrawer()
-  }, [navigation, drawerHelpers])
-
   return (
     <DrawerContentScrollView>
-      <View style={styles.header}>
-        <TouchableOpacity onPress={handlePressAccount}>
-          <ProfilePicture profile={accountUser} style={styles.accountImage} />
-          <View style={styles.accountName}>
-            <Text numberOfLines={1} style={styles.name} variant='h1' noGutter>
-              {name}
-            </Text>
-            <UserBadges
-              user={accountUser}
-              hideName
-              style={styles.accountBadges}
-            />
-          </View>
-          <Text
-            style={styles.handle}
-            numberOfLines={1}
-            weight='medium'
-            fontSize='medium'
-          >
-            @{handle}
-          </Text>
-        </TouchableOpacity>
-        <TouchableOpacity style={styles.tokens} onPress={handlePressRewards}>
-          <IconAudioBadge
-            tier={tier}
-            showNoTier
-            style={styles.tokenBadge}
-            height={spacing(7)}
-            width={spacing(7)}
-          />
-          <Text fontSize='large' weight='heavy'>
-            {formatWei(totalBalance, true, 0)}
-          </Text>
-        </TouchableOpacity>
-      </View>
+      <AccountDetails />
       <VanityMetrics />
       <LeftNavLink
         icon={IconUser}
